# manage-api/openapi.json
{"openapi": "3.0.2", "info": {"title": "FastAPI", "version": "0.1.0"}, "paths": {"/v1/stream": {"get": {"tags": ["stream"], "summary": "Get Stream Apps", "description": "Get all streaming applications.\nstart: Start index of the applications\nsize: Number of sessions to fetch", "operationId": "get_stream_apps_v1_stream_get", "parameters": [{"required": false, "schema": {"title": "Start", "type": "integer", "default": 0}, "name": "start", "in": "query"}, {"required": false, "schema": {"title": "Size", "type": "integer", "default": 10}, "name": "size", "in": "query"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/LivyGetBatchResp"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "post": {"tags": ["stream"], "summary": "Create Stream App", "description": "Create a streaming application through apache livy.", "operationId": "create_stream_app_v1_stream_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/CreateStreamAppReq"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/LivyBatch"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/v1/stream/{id}": {"get": {"tags": ["stream"], "summary": "Get Stream App", "description": "Get detail for a streaming application with {id}", "operationId": "get_stream_app_v1_stream__id__get", "parameters": [{"required": true, "schema": {"title": "Id", "type": "integer"}, "name": "id", "in": "path"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/LivyBatch"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "delete": {"tags": ["stream"], "summary": "Delete Stream App", "description": "Delete a streaming application with {id}", "operationId": "delete_stream_app_v1_stream__id__delete", "parameters": [{"required": true, "schema": {"title": "Id", "type": "integer"}, "name": "id", "in": "path"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/healthcheck": {"get": {"tags": ["healthcheck"], "summary": "Read Healthcheck", "description": "Health check.", "operationId": "read_healthcheck_healthcheck_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}}, "components": {"schemas": {"CreateStreamAppReq": {"title": "CreateStreamAppReq", "type": "object", "properties": {"sparkAppName": {"title": "Sparkappname", "type": "string", "default": ""}}, "description": "Request body for POST /stream"}, "HTTPValidationError": {"title": "HTTPValidationError", "type": "object", "properties": {"detail": {"title": "Detail", "type": "array", "items": {"$ref": "#/components/schemas/ValidationError"}}}}, "LivyBatch": {"title": "LivyBatch", "required": ["id", "log", "state"], "type": "object", "properties": {"id": {"title": "Id", "type": "integer"}, "name": {"title": "Name", "type": "string"}, "appId": {"title": "Appid", "type": "string"}, "log": {"title": "Log", "type": "array", "items": {"type": "string"}}, "state": {"title": "State", "type": "string"}}}, "LivyGetBatchResp": {"title": "LivyGetBatchResp", "required": ["start", "total", "sessions"], "type": "object", "properties": {"start": {"title": "Start", "type": "integer"}, "total": {"title": "Total", "type": "integer"}, "sessions": {"title": "Sessions", "type": "array", "items": {"$ref": "#/components/schemas/LivyBatch"}}}}, "ValidationError": {"title": "ValidationError", "required": ["loc", "msg", "type"], "type": "object", "properties": {"loc": {"title": "Location", "type": "array", "items": {"type": "string"}}, "msg": {"title": "Message", "type": "string"}, "type": {"title": "Error Type", "type": "string"}}}}}}